#4 dodano historie zmian produktu, ogolna poprawa kodu-zabezbieczen

+++ app/Http/Controllers/CategoryController.php
@@ -4,22 +4,35 @@ namespace App\Http\Controllers;
 
 use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
     public function index()
     {
+        if (!Auth::check() || Auth::user()->role !== 'admin') {
+            return redirect('/home')->with('error', 'Unauthorized access');
+        }
+
         $categories = Category::all();
         return view('categories.index', compact('categories'));
     }
 
     public function create()
     {
+        if (!Auth::check() || Auth::user()->role !== 'admin') {
+            return redirect('/home')->with('error', 'Unauthorized access');
+        }
+
         return view('categories.create');
     }
 
     public function store(Request $request)
     {
+        if (!Auth::check() || Auth::user()->role !== 'admin') {
+            return redirect('/home')->with('error', 'Unauthorized access');
+        }
+
         $request->validate([
             'name' => 'required|string|max:255|unique:categories',
         ]);
@@ -33,11 +46,19 @@ class CategoryController extends Controller
 
     public function edit(Category $category)
     {
+        if (!Auth::check() || Auth::user()->role !== 'admin') {
+            return redirect('/home')->with('error', 'Unauthorized access');
+        }
+
         return view('categories.edit', compact('category'));
     }
 
     public function update(Request $request, Category $category)
     {
+        if (!Auth::check() || Auth::user()->role !== 'admin') {
+            return redirect('/home')->with('error', 'Unauthorized access');
+        }
+
         $request->validate([
             'name' => 'required|string|max:255|unique:categories,name,' . $category->id,
         ]);
@@ -50,16 +71,19 @@ class CategoryController extends Controller
     }
 
     public function destroy($id)
-{
-    $category = Category::findOrFail($id);
+    {
+        if (!Auth::check() || Auth::user()->role !== 'admin') {
+            return redirect('/home')->with('error', 'Unauthorized access');
+        }
 
-    // Ustawienie category_id na null w produktach powiązanych z usuwaną kategorią
-    $category->products()->update(['category_id' => null]);
+        $category = Category::findOrFail($id);
 
-    // Usuń kategorię
-    $category->delete();
+        // Ustawienie category_id na null w produktach powiązanych z usuwaną kategorią
+        $category->products()->update(['category_id' => null]);
 
-    return redirect()->route('categories.index')->with('success', 'Category deleted and products detached.');
-}
+        // Usuń kategorię
+        $category->delete();
 
+        return redirect()->route('categories.index')->with('success', 'Category deleted and products detached.');
+    }
 }
